+++ lutimes.c
@@ -1,8 +1,15 @@
 /*
- * Copyright 2021-2022 Gaël PORTAY
+ * Copyright 2021-2022,2024 Gaël PORTAY
  *
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
+
+#include <stdio.h>
+#include <errno.h>
+
+#include <utime.h>
+
+#include "iamroot.h"
 
 /*
  * Stolen from musl (src/legacy/lutimes.c)
@@ -27,3 +34,13 @@ int lutimes(const char *filename, const struct timeval tv[2])
 	}
 	return utimensat(AT_FDCWD, filename, tv ? times : 0, AT_SYMLINK_NOFOLLOW);
 }
+
+#ifdef __GLIBC__
+#if __TIMESIZE == 32
+#ifdef _LARGEFILE64_SOURCE
+int __lutimes64 (const char *__file, const struct timeval __tvp[2])
+     __THROW __nonnull ((1));
+weak_alias(lutimes, __lutimes64);
+#endif
+#endif
+#endif